requirements.txt
certifi==2020.6.20
chardet==3.0.4
docopt==0.6.2
idna==2.10
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pipreqs==0.4.10
pygal==2.4.0
pyral==1.4.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
XlsxWriter==1.3.0
yarg==0.1.9
cairosvg
